refactor(slack): extract messaging family into SlackMessagingMixin (adapter.py god-file slice R2)
